mononoke: fix blobimport importing a 0 revision only
Summary:
We had a small bug in blobimport. If mononoke repo is empty and blobimport
tries to import a single revision whose rev number is 0 then it will
successfull import it but it will report that no revision were imported
i.e. it will print "didn't import any revision" to stderr and won't update the
manifold latest imported revision marker.

That was because we didn't update max_rev_and_bcs_id if rev is equal to
max_rev. This diff fixes it.

Reviewed By: ahornby

Differential Revision: D25421164

fbshipit-source-id: 639ead0ac326a14051d3a4faba568ecb797857a2