edenscm: fsmonitor: sort blackbox shortlists
Summary:
The integration tests rely on the ordering, but watchman doesn't
define an ordering in its file results, so the current implementation is
implicitly tied to the iteration order of the InMemoryView indices inside
watchman.

In D25414061 that implementation is changed to be derived from the order
in which entries are added to a change journal, making the results more
likely to match the readdir directory iteration order.

Since the blackbox fsmonitor integration tests have a defined ordering in their
output, we need to sort these results here.

Reviewed By: singhsrb

Differential Revision: D25468807

fbshipit-source-id: aa83d82cca0ecdbfd1d4bb1309b9f3c31f6c3136