mononoke: refactor log_commits_to_scribe to forcepushrebase
Summary:
This diff does a small refactoring to hopefully make the code a bit clearer.
Previously we were calling log_commits_to_scribe in force_pushrebase and in
run_pushrebase, so  log_commits_to_scribe was called twice. It didn't mean that
the commits were logged twice but it was hard to understand just by looking at
the code.

Let's make it a bit clearer.

Reviewed By: krallin

Differential Revision: D25555712

fbshipit-source-id: bed9754b1645008846a86da665b6f3f3483f30da