cli: collect all errors while removing the repository
Summary:
On Windows, it's possible that not all files could be removed from the
repository due to some other process holding a reference to it. When that
happens the `edenfsctl rm` operation will fail. Sometimes, instead of failing
with the actual reason for why the removal failed, it throws a cryptic "list
index out of range" error.

The reason was that when the file that can't be removed is actually a
directory, the `errors` list would be empty. Since filtering the folders is a
bit silly, let's not do it.

Reviewed By: fanzeyi

Differential Revision: D25836412

fbshipit-source-id: 36f936ff9d7697dfd2f4c68d4e56bdb18b66b06a