mononoke: add an integration test for megarepo unbinding
Summary:
We are going to write docs on how to unbind repositories and bind them later.
This integration test is a "test for documentation" i.e. it should run the same
steps as we'd run ourselves whenever we need to rebind.

The key part here is that we do a merge of unbound small repo into large repo commits

Reviewed By: ikostia

Differential Revision: D25883873

fbshipit-source-id: fac1d7871e52f4e1aa7f15d32d39f2755b803cd3