mononoke: add sync_commit_and_ancestors command to megarepotool
Summary:
Add a command that will unconditionally rewrite commits from one repo to
another. This is similar to what x-repo lookup thrift api does, however this
command doesn't do a few safety checks (in particular, it allows to rewrite
public commits).

This is going to be used for unbinding procedure.

Reviewed By: ikostia

Differential Revision: D25883874

fbshipit-source-id: a40eda9aa40ef7ad63e2403d956871940dd1601d