mononoke: log client_ip only if source_hostname is empty
Summary:
This might be controversial, so I'd like to hear opinions.
We are running low on retention once again, and there's client_ip and
source_hostname field that have largely duplicated info and together use ~14%
of space.

I suggest to log source_hostname if it exists and client_ip if source_hostname
is not set. This should save ~7% of space since most of the time both of these
fields are set.

Reviewed By: krallin

Differential Revision: D25900993

fbshipit-source-id: f3db59e0dde44a5117059c829df7a1a1c811f641