mononoke: make sure we initialize tunables in derived_data_tailer
Summary:
Previously we weren't initializing tunables at all in derived_data_tailer and
probably in lots of other binaries as well.

This is unfortunate, since derived_data_tailer runs for a long time and it
would be good to be able to control its behaviour.

Let's fix it by using init_mononoke() function.

Reviewed By: ikostia

Differential Revision: D25912156

fbshipit-source-id: eabd1c56120d087a169746077c8a7d36855c2b84