mononoke: allow overriding session class
Summary:
We ran into issues when we write too much data to our blobstore sync queue. The
idea we want to try is to prefer to not write to the queue if puts were
successful.

In the next diff we'll start overriding session class when we derive data.
This diff makes it possible to do so.

Reviewed By: krallin

Differential Revision: D25910463

fbshipit-source-id: 2ead5291acbce4014fbb833cac1dc53a6eb61b13